feat: reworked dungeon1 into dungeon1scene and need to now implement dungeonroom
